Esame BE

+++ database/migrations/2024_07_10_135104_create_richieste_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
             $table->id('RichiestaID');
             $table->string('CognomeNomeRichiedente');
             $table->timestamp('DataInserimentoRichiesta');
-            $table->decimal('Importo', 10, 8);
+            $table->float('Importo');
             $table->integer('NumeroRate');
             $table->timestamps();
         });
